# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Supperjumpin API
  version: 0.1.0
paths:
  /v1/me:
    get:
      operationId: getMe
      summary: Fetch the signed-in Account and Player identity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The internal Account and Player attached to the external auth identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Missing or invalid bearer token.
  /v1/jumps/{jumpId}/evidence-upload-authorizations:
    post:
      operationId: authorizeEvidenceUpload
      summary: Authorize direct Evidence upload for a performer's Planned Jump
      security:
        - bearerAuth: []
      parameters:
        - name: jumpId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentType]
              properties:
                contentType:
                  type: string
      responses:
        "201":
          description: Direct upload target and authorization metadata for the Evidence media.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceUploadAuthorization"
        "400":
          description: Invalid Evidence upload authorization request.
        "401":
          description: Missing or invalid bearer token.
        "403":
          description: The signed-in Player is not the performer of this Planned Jump.
        "404":
          description: Planned Jump not found.
  /v1/jumps/{jumpId}/evidence:
    post:
      operationId: submitEvidence
      summary: Finalize an Evidence record and transition a Planned Jump into a Performed Jump
      security:
        - bearerAuth: []
      parameters:
        - name: jumpId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uploadAuthorizationId, caption]
              properties:
                uploadAuthorizationId:
                  type: string
                caption:
                  type: string
      responses:
        "201":
          description: Backend-owned Evidence record and the updated Performed Jump.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceSubmission"
        "400":
          description: Invalid Evidence submission request.
        "401":
          description: Missing or invalid bearer token.
        "403":
          description: The signed-in Player is not the performer of this Planned Jump.
        "404":
          description: Planned Jump or upload authorization not found.
  /v1/guest-sessions:
    post:
      operationId: createGuestSession
      summary: Create a Guest Judge session for unauthenticated Judgments
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: The created Guest Session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestSession"
  /v1/jumps/{jumpId}/judgment:
    post:
      operationId: submitJudgment
      summary: Submit or edit a Judge's Judgment for a Performed Jump (authenticated or Guest)
      security:
        - bearerAuth: []
      parameters:
        - name: jumpId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [commitment, transgression, creativity, presentation]
              properties:
                guestSessionId:
                  type: string
                  description: Required for Guest Judges. Must not be provided when authenticated.
                commitment:
                  type: integer
                  minimum: 1
                  maximum: 4
                transgression:
                  type: integer
                  minimum: 1
                  maximum: 4
                creativity:
                  type: integer
                  minimum: 1
                  maximum: 4
                presentation:
                  type: integer
                  minimum: 1
                  maximum: 4
      responses:
        "200":
          description: The edited Judgment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Judgment"
        "201":
          description: The created Judgment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Judgment"
        "400":
          description: Invalid Judgment request.
        "401":
          description: Authentication or guestSessionId required.
        "403":
          description: The Judge is not eligible for this Performed Jump (self-judge, cap reached, or grace period active).
        "404":
          description: Performed Jump not found.
        "409":
          description: The Judging Window is closed.
  /v1/opens/{year}/{month}/compute:
    post:
      operationId: computeOpenScores
      summary: Compute Open Final Scores for a completed calendar month
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        "200":
          description: Open scores computed successfully.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
        "400":
          description: Invalid year or month.
        "401":
          description: Missing or invalid bearer token.
        "403":
          description: Compute Open scores not allowed.
        "409":
          description: Open month has not soft-closed yet.
  /v1/feed:
    get:
      operationId: getPublicFeed
      summary: Public Feed of recent Jumps (reverse chronological, cursor paginated)
      security: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Paginated Jump cards for the public Feed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicFeedResponse"
        "500":
          description: The feed could not be loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid cursor or limit parameter.
  /v1/jumps/{jumpId}:
    get:
      operationId: getJumpDetail
      summary: Public Jump Detail — full read view or tombstone
      security: []
      parameters:
        - name: jumpId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: >
            Full Jump detail or tombstone response. Live Jumps (Performed, Judged, etc.)
            return JumpDetail. Removed Jumps return JumpTombstone with no content.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/JumpDetail"
                  - $ref: "#/components/schemas/JumpTombstone"
        "404":
          description: Jump not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    MeResponse:
      type: object
      required: [account, player]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        player:
          $ref: "#/components/schemas/Player"
    Account:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          description: Internal Account identifier, distinct from Supabase subject.
        email:
          type: string
          format: email
    Player:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          description: Internal in-game Player identifier.
        displayName:
          type: string
    Jump:
      type: object
      required: [id, playerId, status, source, destination, food, finalScore]
      properties:
        id:
          type: string
        playerId:
          type: string
        status:
          type: string
          enum: [Performed Jump, Judged Jump, Unjudged Jump, Disqualified Jump, Removed Jump]
        source:
          type: string
        destination:
          type: string
        food:
          type: string
        finalScore:
          anyOf:
            - type: integer
            - type: "null"
    EvidenceUploadAuthorization:
      type: object
      required: [id, jumpId, uploadUrl, uploadMethod, uploadHeaders, mediaObjectKey, expiresAt]
      properties:
        id:
          type: string
        jumpId:
          type: string
        uploadUrl:
          type: string
        uploadMethod:
          type: string
          enum: [PUT]
        uploadHeaders:
          type: object
          additionalProperties:
            type: string
        mediaObjectKey:
          type: string
        expiresAt:
          type: string
          format: date-time
    Evidence:
      type: object
      required: [id, jumpId, caption, mediaObjectKey, createdAt]
      properties:
        id:
          type: string
        jumpId:
          type: string
        caption:
          type: string
        mediaObjectKey:
          type: string
        createdAt:
          type: string
          format: date-time
    EvidenceSubmission:
      type: object
      required: [jump, evidence]
      properties:
        jump:
          $ref: "#/components/schemas/Jump"
        evidence:
          $ref: "#/components/schemas/Evidence"
    Judgment:
      type: object
      required: [id, jumpId, commitment, transgression, creativity, presentation]
      properties:
        id:
          type: string
        jumpId:
          type: string
        playerId:
          type: string
          description: Present for authenticated Player judgments. Empty for Guest judgments.
        guestSessionId:
          type: string
          description: Present for Guest judgments. Empty for authenticated Player judgments.
        provenance:
          type: string
          enum: [public, open, season]
          description: Tracks the context in which the Judgment was submitted.
        commitment:
          type: integer
          minimum: 1
          maximum: 4
        transgression:
          type: integer
          minimum: 1
          maximum: 4
        creativity:
          type: integer
          minimum: 1
          maximum: 4
        presentation:
          type: integer
          minimum: 1
          maximum: 4
    GuestSession:
      type: object
      required: [id]
      properties:
        id:
          type: string
    PublicFeedResponse:
      type: object
      required: [jumps, nextCursor]
      properties:
        jumps:
          type: array
          items:
            $ref: "#/components/schemas/JumpCard"
        nextCursor:
          type: string
          nullable: true
    JumpCard:
      type: object
      required: [id, performerName, performerId, source, destination, food, caption, mediaObjectKey, status, gracePeriodExpiresAt, runningAverage, judgmentCount, createdAt]
      properties:
        id:
          type: string
        performerName:
          type: string
        performerId:
          type: string
        source:
          type: string
        destination:
          type: string
        food:
          type: string
        caption:
          type: string
        mediaObjectKey:
          type: string
        status:
          type: string
          enum: [Performed Jump, Judged Jump, Unjudged Jump, Disqualified Jump]
        gracePeriodExpiresAt:
          type: string
          format: date-time
        runningAverage:
          type: number
          format: float
        judgmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    JumpDetail:
      type: object
      required: [id, performerName, performerId, source, destination, food, caption, mediaObjectKey, status, gracePeriodExpiresAt, runningAverage, judgmentCount, createdAt]
      properties:
        id:
          type: string
        performerName:
          type: string
        performerId:
          type: string
        source:
          type: string
        destination:
          type: string
        food:
          type: string
        caption:
          type: string
        mediaObjectKey:
          type: string
        status:
          type: string
          enum: [Performed Jump, Judged Jump, Unjudged Jump, Disqualified Jump]
        gracePeriodExpiresAt:
          type: string
          format: date-time
        runningAverage:
          type: number
          format: float
        judgmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        finalScore:
          type: integer
          nullable: true
        viewerContext:
          $ref: "#/components/schemas/ViewerContext"
    JumpTombstone:
      type: object
      required: [id, status, message, removedAt]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [Removed Jump]
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        removedAt:
          type: string
          format: date-time
    ViewerContext:
      type: object
      required: [canJudge, hasJudged]
      properties:
        canJudge:
          type: boolean
        reason:
          type: string
          nullable: true
          enum: [self-judging, grace-period, already-judged]
        gracePeriodEndsAt:
          type: string
          format: date-time
          nullable: true
        hasJudged:
          type: boolean